+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import ScriptViewer from './components/script-viewer.js';
-import SCRIPT_MODES from './consts.js';
+import ScriptViewer from './components/script-viewer/script-viewer.js';
+import SCRIPT_MODES from './components/script-viewer/consts.js';
 
 import script from './data/short_replies_annotated.json';
 import { FullStory } from 'react-fullstory-component';
