+++ src/components/image-card-viewer/image-card-viewer.css
