+++ src/stories/index.js
@@ -6,8 +6,8 @@ import { linkTo } from '@storybook/addon-links';
 
 import { Button, Welcome } from '@storybook/react/demo';
 
-import ScriptViewer from '../components/script-viewer.js'
-import SCRIPT_MODES from '../consts.js';
+import ScriptViewer from '../components/script-viewer/script-viewer.js'
+import SCRIPT_MODES from '../components/script-viewer/consts.js';
 
 import script from '../data/short_replies_annotated.json';
